Finished coding in the second data type which is basically counter_counter and everything similar to it of which there are quite a few.
Additionally sorted out the place to save the temp files utilizing serialize and building the scripts to by dynamic in as many ways as 
possible. This includes being able to query all columns not just a specific one and using a custom output handler instead of the 
wmi-include.php function as this would have problems when there are multiple keys of the same name.